+++ cli/src/balance.rs
@@ -19,6 +19,7 @@ pub use self::{cancel::*, pause::*, resume::*, start::*, status::*};
 /// be queried at any time. Requires CAP_SYS_ADMIN.
 #[derive(Parser, Debug)]
 #[allow(clippy::doc_markdown)]
+#[clap(arg_required_else_help = true)]
 pub struct BalanceCommand {
     #[clap(subcommand)]
     pub subcommand: BalanceSubcommand,