+++ cli/src/device.rs
@@ -17,6 +17,7 @@ pub use self::{add::*, ready::*, remove::*, scan::*, stats::*, usage::*};
 /// checking device readiness. Most operations require CAP_SYS_ADMIN.
 #[derive(Parser, Debug)]
 #[allow(clippy::doc_markdown)]
+#[clap(arg_required_else_help = true)]
 pub struct DeviceCommand {
     #[clap(subcommand)]
     pub subcommand: DeviceSubcommand,