+++ cli/src/quota.rs
@@ -15,6 +15,7 @@ pub use self::{disable::*, enable::*, rescan::*, status::*};
 /// subvolume or quota group. Quota operations require CAP_SYS_ADMIN.
 #[derive(Parser, Debug)]
 #[allow(clippy::doc_markdown)]
+#[clap(arg_required_else_help = true)]
 pub struct QuotaCommand {
     #[clap(subcommand)]
     pub subcommand: QuotaSubcommand,