+++ cli/src/subvolume.rs
@@ -25,6 +25,7 @@ pub use self::{
 /// Most operations require CAP_SYS_ADMIN.
 #[derive(Parser, Debug)]
 #[allow(clippy::doc_markdown)]
+#[clap(arg_required_else_help = true)]
 pub struct SubvolumeCommand {
     #[clap(subcommand)]
     pub subcommand: SubvolumeSubcommand,